+++ packages/effect/src/Inspectable.ts
@@ -40,7 +40,7 @@ export const toJSON = (x: unknown): unknown => {
     } else if (Array.isArray(x)) {
       return x.map(toJSON)
     }
-  } catch (_) {
+  } catch {
     return {}
   }
   return redact(x)
@@ -97,7 +97,7 @@ export const toStringUnknown = (u: unknown, whitespace: number | string | undefi
   }
   try {
     return typeof u === "object" ? stringifyCircular(u, whitespace) : String(u)
-  } catch (_) {
+  } catch {
     return String(u)
   }
 }