+++ packages/effect/test/RedBlackTree.test.ts
@@ -42,7 +42,7 @@ describe("RedBlackTree", () => {
     if (typeof window !== "undefined") {
       return
     }
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    // eslint-disable-next-line @typescript-eslint/no-require-imports
     const { inspect } = require("node:util")
 
     const tree = pipe(