+++ packages/experimental/examples/machine.ts
@@ -34,7 +34,7 @@ const mailer = Machine.makeWith<List.List<SendEmail>>()((_, previous) =>
     const state = previous ?? List.empty()
 
     if (List.isCons(state)) {
-      yield* ctx.unsafeSend(new ProcessEmail()), Effect.replicateEffect(List.size(state))
+      yield* ctx.unsafeSend(new ProcessEmail()).pipe(Effect.replicateEffect(List.size(state)))
     }
 
     return Machine.procedures.make(state).pipe(
@@ -44,7 +44,7 @@ const mailer = Machine.makeWith<List.List<SendEmail>>()((_, previous) =>
             return [void 0, state]
           }
           const req = state.head
-          yield* pipe(Effect.log(`Sending email to ${req.email}`), Effect.delay(500))
+          yield* Effect.log(`Sending email to ${req.email}`).pipe(Effect.delay(500))
           return [void 0, state.tail]
         })),
       Machine.procedures.add<SendEmail>()("SendEmail", (ctx) =>