+++ packages/opentelemetry/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @effect/opentelemetry
 
+## 0.48.3
+
+### Patch Changes
+
+- Updated dependencies [[`b8722b8`](https://github.com/Effect-TS/effect/commit/b8722b817e2306fe8c8245f3f9e32d85b824b961), [`a328f4b`](https://github.com/Effect-TS/effect/commit/a328f4b4fe717dd53e5b04a30f387433c32f7328)]:
+  - effect@3.15.2
+  - @effect/platform@0.82.3
+
+## 0.48.2
+
+### Patch Changes
+
+- Updated dependencies [[`739a3d4`](https://github.com/Effect-TS/effect/commit/739a3d4a4565915fe2e690003f4f9085cb4422fc)]:
+  - @effect/platform@0.82.2
+
 ## 0.48.1
 
 ### Patch Changes